refactoring common methods/logic into helper modules
